# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CardWithMemo": {
        "additionalProperties": false,
        "properties": {
          "card": {
            "$ref": "#/components/schemas/TradingIdeaCard"
          },
          "memo": {
            "$ref": "#/components/schemas/ResearchMemo"
          }
        },
        "required": [
          "card",
          "memo"
        ],
        "title": "CardWithMemo",
        "type": "object"
      },
...
      "ExtractionRequestRecord": {
        "additionalProperties": false,
        "properties": {
          "extraction_config_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Extraction Config Sha256",
            "type": "string"
          },
          "extraction_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Model Id"
          },
          "extraction_model_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Model Revision"
          },
          "extraction_prompt_sha256": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Prompt Sha256"
          },
          "extraction_prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Prompt Version"
          },
          "extraction_request_id": {
            "format": "uuid",
            "title": "Extraction Request Id",
            "type": "string"
          },
          "extraction_schema_version": {
            "default": "phase2-trading-idea-card-v2",
            "title": "Extraction Schema Version",
            "type": "string"
          },
          "extractor_id": {
            "default": "fable5-deterministic-extractor",
            "title": "Extractor Id",
            "type": "string"
          },
          "extractor_kind": {
            "$ref": "#/components/schemas/ExtractorKind",
            "default": "deterministic_mock"
          },
          "extractor_version": {
            "default": "1",
            "title": "Extractor Version",
            "type": "string"
          },
          "latest_event": {
            "$ref": "#/components/schemas/ExtractionEventType"
          },
          "request_fingerprint": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Request Fingerprint",
            "type": "string"
          },
          "requested_at_utc": {
            "format": "date-time",
            "title": "Requested At Utc",
            "type": "string"
          },
          "rq_job_id": {
            "title": "Rq Job Id",
            "type": "string"
          },
          "source_version_id": {
            "format": "uuid",
            "title": "Source Version Id",
            "type": "string"
          }
        },
        "required": [
          "extraction_config_sha256",
          "extraction_request_id",
          "source_version_id",
          "request_fingerprint",
          "rq_job_id",
          "latest_event",
          "requested_at_utc"
        ],
        "title": "ExtractionRequestRecord",
        "type": "object"
      },
      "ExtractorKind": {
        "enum": [
          "deterministic_mock",
          "llm"
        ],
        "title": "ExtractorKind",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResearchMemo": {
        "additionalProperties": false,
        "properties": {
          "card_id": {
            "format": "uuid",
            "title": "Card Id",
            "type": "string"
          },
          "content_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Content Sha256",
            "type": "string"
          },
          "created_at_utc": {
            "format": "date-time",
            "title": "Created At Utc",
            "type": "string"
          },
          "markdown": {
            "title": "Markdown",
            "type": "string"
          },
          "memo_id": {
            "format": "uuid",
            "title": "Memo Id",
            "type": "string"
          },
          "template_version": {
            "const": "phase2-memo-v1",
            "default": "phase2-memo-v1",
            "title": "Template Version",
            "type": "string"
          }
        },
        "required": [
          "memo_id",
          "card_id",
          "markdown",
          "content_sha256",
          "created_at_utc"
        ],
        "title": "ResearchMemo",
        "type": "object"
      },
...
      "SourceVersion": {
        "additionalProperties": false,
        "properties": {
          "authority_verification_method": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthorityVerificationMethod"
              },
              {
                "type": "null"
              }
            ]
          },
          "content_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Content Sha256",
            "type": "string"
          },
          "content_state": {
            "$ref": "#/components/schemas/ContentState"
          },
          "created_at_utc": {
            "format": "date-time",
            "title": "Created At Utc",
            "type": "string"
          },
          "official_corroboration_source_version_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Official Corroboration Source Version Ids",
            "type": "array"
          },
          "parent_source_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Source Version Id"
          },
          "raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text"
          },
          "retrieved_at_utc": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieved At Utc"
          },
          "source_authority": {
            "$ref": "#/components/schemas/SourceAuthority"
          },
          "source_id": {
            "format": "uuid",
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "source_version": {
            "minimum": 1.0,
            "title": "Source Version",
            "type": "integer"
          },
          "source_version_id": {
            "format": "uuid",
            "title": "Source Version Id",
            "type": "string"
          },
          "supplied_at_utc": {
            "format": "date-time",
            "title": "Supplied At Utc",
            "type": "string"
          }
        },
        "required": [
          "source_version_id",
          "source_id",
          "source_version",
          "parent_source_version_id",
          "source_type",
          "source_authority",
          "source_url",
          "content_state",
          "raw_text",
          "content_sha256",
          "supplied_at_utc",
          "retrieved_at_utc",
          "authority_verification_method",
          "created_at_utc"
        ],
        "title": "SourceVersion",
        "type": "object"
      },
...
      "TradingIdeaCard": {
        "additionalProperties": false,
        "properties": {
          "action_rule": {
            "$ref": "#/components/schemas/ActionRuleEvidence"
          },
          "ambiguity_flags": {
            "items": {
              "$ref": "#/components/schemas/AmbiguityFlag"
            },
            "title": "Ambiguity Flags",
            "type": "array"
          },
          "asset_class": {
            "$ref": "#/components/schemas/AssetClassEvidence"
          },
          "card_id": {
            "format": "uuid",
            "title": "Card Id",
            "type": "string"
          },
          "contribution_status": {
            "$ref": "#/components/schemas/ContributionStatus"
          },
          "corroboration_status": {
            "$ref": "#/components/schemas/CorroborationStatus"
          },
          "created_at_utc": {
            "format": "date-time",
            "title": "Created At Utc",
            "type": "string"
          },
          "execution_style": {
            "$ref": "#/components/schemas/ExecutionStyleEvidence"
          },
          "extraction_config_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Extraction Config Sha256",
            "type": "string"
          },
          "extraction_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Model Id"
          },
          "extraction_model_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Model Revision"
          },
          "extraction_prompt_sha256": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Prompt Sha256"
          },
          "extraction_prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Prompt Version"
          },
          "extraction_request_id": {
            "format": "uuid",
            "title": "Extraction Request Id",
            "type": "string"
          },
          "extraction_schema_version": {
            "title": "Extraction Schema Version",
            "type": "string"
          },
          "extractor_id": {
            "title": "Extractor Id",
            "type": "string"
          },
          "extractor_kind": {
            "$ref": "#/components/schemas/ExtractorKind"
          },
          "extractor_version": {
            "title": "Extractor Version",
            "type": "string"
          },
          "forecast_horizon": {
            "$ref": "#/components/schemas/ForecastHorizonEvidence"
          },
          "infra_risk": {
            "$ref": "#/components/schemas/InfraRisk"
          },
          "official_corroboration_source_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Official Corroboration Source Ids",
            "type": "array"
          },
          "official_corroboration_source_version_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Official Corroboration Source Version Ids",
            "type": "array"
          },
          "paraphrased_claim": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paraphrased Claim"
          },
          "quoted_claims": {
            "items": {
              "$ref": "#/components/schemas/QuotedClaim"
            },
            "title": "Quoted Claims",
            "type": "array"
          },
          "raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text"
          },
          "required_data": {
            "$ref": "#/components/schemas/RequiredDataEvidence"
          },
          "research_priority_score": {
            "title": "Research Priority Score",
            "type": "null"
          },
          "risk_assumptions": {
            "$ref": "#/components/schemas/RiskAssumptionsEvidence"
          },
          "signal_family": {
            "$ref": "#/components/schemas/SignalFamilyEvidence"
          },
          "source_authority": {
            "$ref": "#/components/schemas/SourceAuthority"
          },
          "source_id": {
            "format": "uuid",
            "title": "Source Id",
            "type": "string"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "source_version": {
            "minimum": 1.0,
            "title": "Source Version",
            "type": "integer"
          },
          "source_version_id": {
            "format": "uuid",
            "title": "Source Version Id",
            "type": "string"
          },
          "synthetic_fixture": {
            "default": false,
            "title": "Synthetic Fixture",
            "type": "boolean"
          },
          "testability_reason_codes": {
            "items": {
              "$ref": "#/components/schemas/TestabilityReason"
            },
            "title": "Testability Reason Codes",
            "type": "array"
          },
          "testability_score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Testability Score",
            "type": "number"
          },
          "testability_score_method": {
            "const": "phase2-testability-v1",
            "default": "phase2-testability-v1",
            "title": "Testability Score Method",
            "type": "string"
          },
          "testability_status": {
            "$ref": "#/components/schemas/TestabilityStatus"
          }
        },
        "required": [
          "card_id",
          "extraction_request_id",
          "source_id",
          "source_version_id",
          "source_authority",
          "source_url",
          "source_version",
          "raw_text",
          "quoted_claims",
          "paraphrased_claim",
          "asset_class",
          "forecast_horizon",
          "signal_family",
          "execution_style",
          "required_data",
          "action_rule",
          "risk_assumptions",
          "ambiguity_flags",
          "testability_status",
          "testability_reason_codes",
          "testability_score",
          "infra_risk",
          "research_priority_score",
          "corroboration_status",
          "contribution_status",
          "official_corroboration_source_ids",
          "official_corroboration_source_version_ids",
          "extractor_kind",
          "extractor_id",
          "extractor_version",
          "extraction_model_id",
          "extraction_model_revision",
          "extraction_prompt_version",
          "extraction_prompt_sha256",
          "extraction_schema_version",
          "extraction_config_sha256",
          "created_at_utc"
        ],
        "title": "TradingIdeaCard",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Research-to-paper-trading platform API. No live execution path exists.",
    "title": "Fable5 Research API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/cards": {
      "get": {
        "operationId": "list_cards_v1_cards_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TradingIdeaCard"
                  },
                  "title": "Response List Cards V1 Cards Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Cards",
        "tags": [
          "idea-intake"
        ]
      }
    },
    "/v1/cards/{card_id}": {
      "get": {
        "operationId": "get_card_v1_cards__card_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "card_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Card Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardWithMemo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Card",
        "tags": [
          "idea-intake"
        ]
      }
    },
...
    "/v1/cards/{card_id}/memo": {
      "get": {
        "operationId": "get_memo_v1_cards__card_id__memo_get",
        "parameters": [
          {
            "in": "path",
            "name": "card_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Card Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchMemo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Memo",
        "tags": [
          "idea-intake"
        ]
      }
    },
    "/v1/extractions": {
      "get": {
        "operationId": "list_extractions_v1_extractions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExtractionRequestRecord"
                  },
                  "title": "Response List Extractions V1 Extractions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Extractions",
        "tags": [
          "idea-intake"
        ]
      }
    },
    "/v1/extractions/{request_id}": {
      "get": {
        "operationId": "get_extraction_v1_extractions__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionRequestRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Extraction",
        "tags": [
          "idea-intake"
        ]
      }
    },
...
    "/v1/source-versions/{source_version_id}": {
      "get": {
        "operationId": "get_source_version_v1_source_versions__source_version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "source_version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceVersion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Source Version",
        "tags": [
          "idea-intake"
        ]
      }
    },
    "/v1/source-versions/{source_version_id}/extractions": {
      "post": {
        "operationId": "create_extraction_v1_source_versions__source_version_id__extractions_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionRequestRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Extraction",
        "tags": [
          "idea-intake"
        ]
      }
    },
    "/v1/sources": {
      "get": {
        "operationId": "list_sources_v1_sources_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceRecord"
                  },
                  "title": "Response List Sources V1 Sources Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sources",
        "tags": [
          "idea-intake"
        ]
      },
      "post": {
        "operationId": "create_source_v1_sources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceIntakeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Source",
        "tags": [
          "idea-intake"
        ]
      }
    },
    "/v1/sources/{source_id}": {
      "get": {
        "operationId": "get_source_v1_sources__source_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Source",
        "tags": [
          "idea-intake"
        ]
      }
    },
    "/v1/sources/{source_id}/versions": {
      "post": {
        "operationId": "add_source_version_v1_sources__source_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceCorrectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Source Version",
        "tags": [
          "idea-intake"
        ]
      }
    }
  }
}